EW-497 working on review comments
