+++ config/config.prod.json
@@ -3,7 +3,7 @@
         "PORT": 8080
     },
     "DB": {
-        "CLIENT_URL": "postgres://admin:password@127.0.0.1:5432",
+        "CLIENT_URL": "postgres://127.0.0.1:5432",
         "DB_NAME": "dbildungs-iam"
     }
 }
