+++ src/modules/person/api/person.enums.ts
@@ -1,0 +1,13 @@
+export enum PersonGender {
+    MALE = 'm',
+    FEMALE = 'f',
+    DIVERSE = 'd',
+    UNKNOWN = 'x',
+}
+
+export enum PersonTrustLevel {
+    NONE = 'kein',
+    UNKNOWN = 'unbe',
+    TRUSTED = 'teil',
+    VERIFIED = 'voll',
+}
