+++ src/modules/person/domain/person.enums.ts
@@ -1,22 +1,8 @@
-export enum PersonGender {
-    MALE = 'm',
-    FEMALE = 'f',
-    DIVERSE = 'd',
-    UNKNOWN = 'x',
-}
-
 export enum Gender {
     MALE = 'male',
     FEMALE = 'female',
     DIVERSE = 'diverse',
     UNKNOWN = 'unknown',
-}
-
-export enum PersonTrustLevel {
-    NONE = 'kein',
-    UNKNOWN = 'unbe',
-    TRUSTED = 'teil',
-    VERIFIED = 'voll',
 }
 
 export enum TrustLevel {
